Opiton for t5 tasks to use a combined preprocessor (instead of separate text preprocessors and token preprocessors).  Combined preprocessors are not compatible with task caching.
Minor hack: Negative target ids now indicate tokens that are to be used as
autoregressive inputs, but not counted in the loss.
For examples, target ids [-4, -8, 3, 1] lead to autoregressive inputs
[0, 4, 8, 3], but there would only be losses on the last two predicted tokens.
The more principled way to do this in the future will be for the task to
generate an additional targets_weights tensor.

PiperOrigin-RevId: 330824009